不同用户收费不同
基本已经完成

+++ read-P/read-commons/src/main/java/com/yong/springcloud/entities/Author.java
@@ -1,6 +1,7 @@
 package com.yong.springcloud.entities;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -15,7 +16,7 @@ public class Author implements Serializable{
 	private String authorname;
 	private String authoremail;
 	private String authorsex;
-	private Long authorincome;
+	private BigDecimal authorincome;
 	private String authorintroduction;
 	private String authorworks;
 	private Long authordots;
